Reorganized CRA files in new folder structure.
